Moved data labels for column and pie in each series class.
